requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
databases==0.9.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
